// src/app/models/strategicData.ts
export interface StrategicData {
  id: number;
  totalStockValue: number;
  totalSetupEvents: number;
  totalWageIdleTimeCosts: number;
  totalWageCosts: number;
  totalMachineIdleTimeCosts: number;
  normalCapacityCurrent: number;
  normalCapacityAverage: number;
  possibleCapacityCurrent: number;
  possibleCapacityAverage: number;
  relPossibleNormalCapacityCurrent: number;
  relPossibleNormalCapacityAverage: number;
  productiveTimeCurrent: number;
  productiveTimeAverage: number;
  effiencyCurrent: number;
  effiencyAverage: number;
  sellwishCurrent: number;
  sellwishAverage: number;
  salesQuantityCurrent: number;
  salesQuantityAverage: number;
  deliveryReliabilityCurrent: number;
  deliveryReliabilityAverage: number;
  idleTimeCurrent: number;
  idleTimeAverage: number;
  idleTimeCostsCurrent: number;
  idleTimeCostsAverage: number;
  storeValueCurrent: number;
  storeValueAverage: number;
  storageCostsCurrent: number;
  storageCostsAverage: number;
  quantityDefectiveGoodsCurrent: number;
  quantityDefectiveGoodsAverage: number;
  costsDefectiveGoodsCurrent: number;
  costsDefectiveGoodsAverage: number;
  normalSalesPriceCurrent: number;
  normalSalesPriceAverage: number;
  normalSaleProfitCurrent: number;
  normalSaleProfitAverage: number;
  directSaleProfitCurrent: number;
  directSaleProfitAverage: number;
  directSaleContractPenaltyCurrent: number;
  directSaleContractPenaltyAverage: number;
  marketplaceSaleProfitCurrent: number;
  marketplaceSaleProfitAverage: number;
  totalProfitCurrent: number;
  totalProfitAverage: number;
}